Add puppet-tripleo module.
As much as possible we are trying to avoid composition layers with
our TripleO puppet integration and just use the stackforge
puppet-* modules directly. There are some exceptions to this
for things like setting up the loadbalancers, and databases...
so we are starting a puppet-tripleo project.

For now puppet-triplo lives on github. Once we have approval the hope
is it will live alongside tripleo-puppet-elements on git.openstack.org
under TripleO.

Change-Id: I04b9e40cdf521c91d00f21fa36b07aa1fb3057f8