Install nfs-utils on compute nodes
The package is required on compute nodes when using NFS backend for
Cinder. This won't make a difference between working and non-working
cloud now that we build a joint overcloud image for all
roles (controller element will pull in nfs-utils). However, for
completeness sake the nfs-utils package should be listed in the compute
element too.

Change-Id: Ia08dac977a2ad2d5f142223561e714d1b4e2fca5