Support for the Opencontrail controller
This commit aims to add the neccesary puppet bits so one can deploy an
opencontrail controller.

The puppet-contrail modules aims to configure all aspects of contrail on
both controller and vrouter.

The puppet-cassandra, puppet-kafka and puppet-zookeeper modules are here
in order to configure properly the respective databases on which
Opencontrail controller relies.

Change-Id: I1b0f5b4ae4a892c687007e79e3ffd60851552879
Depends-On: I7455c71ca8747ee05b5a87362e2bd6a5527f86c8