Use package-installs in puppet-modules element
Migrates the puppet-modules element to use the package-installs element
instead of calling install-packages directly.

This helps us remove additional element dependencies in the undercloud
install by not having to rely on the yum element which provides
install-packages.

This change along with an instack-undercloud packaging change:
https://review.rdoproject.org/r/#/c/1206/ makes it such that the
undercloud install does not have to depend on the yum element, whether
it's using package or source for puppet-modules.

Change-Id: Ia8fb1b8d356d578adc58d0a53a573a39de024e9f
Closes-Bug: 1584864