Move nova.conf truncation into separate element
This patch refactors the hiera elements 40-truncate-nova-config
os-refresh-config script out into a separate element. I recently
tried deploying without this truncation and it still appears
to be required. As we are moving towards using a different
hiera element we need to include this element separately.

Note: The new hiera agent heat hook is required before we can use
t-h-t in the undercloud so resolving this issue is a prerequisite
for that effort.

Change-Id: I6b2da9fbb6cf418a388d29db5148087b33f79b4f
Partially-Implements: blueprint heat-undercloud