Install docker in overcloud-base element
The heat templates for containerized compute have merged, but they
currently expect the host image to be atomic, not overcloud-full.

Docker needs to be installed on overcloud-full so that
containerized compute can switch to overcloud-full, and work can start
on the templates for containerized controller.

Enabling and starting docker will need to be done via some boot-time
configuration.

Change-Id: I1eab2a6de721c8f3c21c7df0019f2d4d1cc3775f