Add ansible-pacemaker to the overcloud image
This is a requirement of major upgrades, and soon will probably become
a requirement of minor updates too. Having it in fresh deployments
from Pike will allow us to use it for minor updates. Even if we only
needed it for major upgrades still, having it in fresh Pike
deployments will allow us to drop the code that would explicitly
install it during P->Q upgrade.

Co-Authored-By: Marius Cornea <mcornea@redhat.com>
Change-Id: Ic8ce72133c47a4c90d581a0925213877b11a471e
Partial-Bug: #1701208