Create element to purge resolv.conf
We should be managing the network configuration with os-net-config or
elsewhere so we should not be providing a nameserver in resolv.conf
in overcloud images. This element will be used to ensure that the
resolv.conf of the overcloud-full image should not contain any
nameservers.

Change-Id: If8c44137d120059b356675a074e3af23ea63a6f8
Partial-Bug: #1749756
(cherry picked from commit c38d7ff551673d74b3e15415939555e8258936b5)