Remove Openstack and Ceph packages attempt 2
As we've switched to containers, it's time to start cleaning up the
packages we install on the overcloud images by default. This change
removes all the openstack and ceph packages from the
overcloud-controller, overcloud-ceph-storage and overcloud-compute
elements.

This reverts commit c69d6dd2c9d1c54d82669525bababc04b5a0688c. The
previous attempt was reverted due to an invalid json in the overcloud
pkg-map. This trailing comma has been addressed in this revert.

Change-Id: I5ceb329e18024040697ad5eab892d13c51ad276f
Related-Blueprint: cleanup-overcloud-images