Port items from instack-undercloud
Currently we leverage the overcloud-full and pip-and-virtualenv-override
elements from instack-undercloud during image building. In order to
remove instack-undercloud as a dependency, we need to move these items
elsewhere.  Since you cannot have overlapping names, the overcloud-full
element functionality has been moved to overcloud-base and
overcloud-compute with the file names 50 being replaced with 51.
The pip-and-virtualenv-override elements has been renamed to
override-pip-and-virtualenv.  A subsequent set of patches will be posted
to tripleo-common to remove instack-undercloud from the image building
path and an update to the elements required when building images.

Change-Id: I1de0b6421ef455ef58c0146da7fe52c35682b395
Related-Blueprint: remove-instack-undercloud