Include cephadm in the overcloud image
As described in the spec [1], ceph-ansible is going to be
deprecated in favour of cephadm and it's distributed as an
RPM that can be installed from Ceph repositories.
However, since the deployment approach is changed and cephadm
requires a Ceph monitor node to bootstrap a minimal cluster,
we should include cephadm RPM on the overcloud image.
This patch adds the cephadm package to the overcloud image.

[1] https://review.opendev.org/c/openstack/tripleo-specs/+/723108

Depends-On: If56c388d8283c93ae8bcb2af3e775d8f6d19b704
Change-Id: I1211e8ac41fdcedf8c64c17936376e1878c467a4