Add nftables to the OC images
In order to properly switch the firewall to nftables, we of course need
its actual package to be installed. Currently, only iptables-nft wrapper
is present - this includes the `nft' binary, but not the service.

Note that we can live with both binaries - this is needed anyway, since
services such as neutron are still calling `iptables' directly.

Change-Id: Ie2443668fea175537b500da6f70e25ab82f59a46